Upgrade to v0.28.0

+++ src/activitygen/activities/AGWorkAndSchool.cpp
@@ -5,7 +5,7 @@
 /// @author  Michael Behrisch
 /// @author  Walter Bamberger
 /// @date    July 2010
-/// @version $Id: AGWorkAndSchool.cpp 20433 2016-04-13 08:00:14Z behrisch $
+/// @version $Id: AGWorkAndSchool.cpp 21329 2016-08-24 12:47:50Z behrisch $
 ///
 // Generates trips to work and to school
 /****************************************************************************/
@@ -137,7 +137,7 @@ AGWorkAndSchool::carAllocation() {
     }
     if (personsDrivingCars.empty() && ! childrenNeedingCarAccompaniment.empty()) {
         //at least one adult exists because no household contains less than one adult
-        if (workingPeoplePossCar.size() != myHousehold->getAdultNbr()) { //personsDrivingCars.size() + adultNeedingCarAccompaniment.size() is equal to 0
+        if ((int)workingPeoplePossCar.size() != myHousehold->getAdultNbr()) { //personsDrivingCars.size() + adultNeedingCarAccompaniment.size() is equal to 0
             std::list<AGAdult>::const_iterator itUA;
             for (itUA = myHousehold->getAdults().begin(); itUA != myHousehold->getAdults().end(); ++itUA) {
                 if (! itUA->isWorking()) {
@@ -374,7 +374,7 @@ AGWorkAndSchool::getUnusedCar() {
 void
 AGWorkAndSchool::makePossibleDriversDrive() {
     //give to a non working adult the ability to drive children or someone else.
-    if (workingPeoplePossCar.size() + personsDrivingCars.size() + adultNeedingCarAccompaniment.size() != myHousehold->getAdultNbr()) {
+    if ((int)(workingPeoplePossCar.size() + personsDrivingCars.size() + adultNeedingCarAccompaniment.size()) != myHousehold->getAdultNbr()) {
         std::list<AGAdult>::const_iterator itUA;
         for (itUA = myHousehold->getAdults().begin(); itUA != myHousehold->getAdults().end(); ++itUA) {
             if (! itUA->isWorking()) {
